refactor: dogs 관련 query 훅 이동
